Module Calculon__Plugin_history

History plugin

val plugin : ?⁠default_len:int -> ?⁠n:int -> unit -> Calculon.Plugin.t

plugin for storing n lines of history and giving them back in query to people who ask.

parameter n

the number of lines to retain internally

parameter default_len

the number of lines to display by default, when replying to message "!history"